requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
colorama==0.4.5
colorlog==6.7.0
colormap==1.0.4
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
Deprecated==1.2.13
Django==4.1.1
django-leaflet==0.28.3
djangorestframework==3.14.0
easydev==0.12.0
et-xmlfile==1.1.0
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.27
httplib2==0.20.4
idna==3.4
jdcal==1.4.1
kiwisolver==1.4.4
leaflet==0.0.3
matplotlib==3.6.0
numpy==1.23.3
ogr==0.40.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pexpect==4.8.0
Pillow==9.2.0
psycopg2==2.9.3
ptyprocess==0.7.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.5.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-gitlab==3.10.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
urllib3==1.26.12
wget==3.2
wrapt==1.14.1

